Fixed compiler warning in beam_shared: implicit conversion from 'double' to 'int' changes value from 25.6 to 25 [-Wliteral-conversion]
Fixed on the assumption that a float variable with float getter should also have a float setter, not an int setter.

Building on Mac OS X 10.8.4, Apple LLVM version 4.2 (clang-425.0.28) (based on LLVM 3.2svn)

REDACTED/mp/src/game/server/hl2mp/grenade_tripmine.cpp:179:26: warning: implicit conversion from 'double' to 'int' changes value from 25.6 to 25 [-Wliteral-conversion]
 m_pBeam->SetScrollRate( 25.6 );
 ~~~~~~~                 ^~~~

+++ sp/src/game/shared/beam_shared.h
@@ -101,7 +101,7 @@ public:
 	void SetColor( int r, int g, int b );
 	void SetBrightness( int brightness );
 	void SetFrame( float frame );
-	void SetScrollRate( int speed );
+	void SetScrollRate( float speed );
 	void SetFireTime( float flFireTime );
 	void SetFrameRate( float flFrameRate ) { m_flFrameRate = flFrameRate; }
 
@@ -351,7 +351,7 @@ inline void CBeam::SetFrame( float frame )
 	m_fStartFrame = frame; 
 }
 
-inline void CBeam::SetScrollRate( int speed )			
+inline void CBeam::SetScrollRate( float speed )
 { 
 	m_fSpeed = speed; 
 }
